requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
groq
openai
pillow
requests
python-multipart
typing
PyPDF2
